+++ modules/alloy/src/main/ts/ephox/alloy/api/behaviour/Positioning.ts
@@ -1,5 +1,6 @@
 import * as ActivePosition from '../../behaviour/positioning/ActivePosition';
 import * as PositionApis from '../../behaviour/positioning/PositionApis';
+import * as PositioningState from '../../behaviour/positioning/PositioningState';
 import { PositioningBehaviour } from '../../behaviour/positioning/PositioningTypes';
 import { PositionSchema } from '../../behaviour/positioning/PositionSchema';
 import * as Behaviour from './Behaviour';
@@ -8,7 +9,8 @@ const Positioning: PositioningBehaviour = Behaviour.create({
   fields: PositionSchema,
   name: 'positioning',
   active: ActivePosition,
-  apis: PositionApis
+  apis: PositionApis,
+  state: PositioningState
 });
 
 export {