+++ modules/alloy/src/main/ts/ephox/alloy/positioning/view/PositionCss.ts
@@ -3,10 +3,10 @@ import { Css, SugarElement } from '@ephox/sugar';
 
 export interface PositionCss {
   readonly position: string;
-  readonly left: Optional<number>;
-  readonly top: Optional<number>;
-  readonly right: Optional<number>;
-  readonly bottom: Optional<number>;
+  readonly left: Optional<string>;
+  readonly top: Optional<string>;
+  readonly right: Optional<string>;
+  readonly bottom: Optional<string>;
 }
 
 const NuPositionCss = (
@@ -15,24 +15,24 @@ const NuPositionCss = (
   top: Optional<number>,
   right: Optional<number>,
   bottom: Optional<number>
-): PositionCss => ({
-  position,
-  left,
-  top,
-  right,
-  bottom
+): PositionCss => {
+  const toPx = (num: number) => num + 'px';
+  return {
+    position,
+    left: left.map(toPx),
+    top: top.map(toPx),
+    right: right.map(toPx),
+    bottom: bottom.map(toPx)
+  };
+};
+
+const toOptions = (position: PositionCss): Record<string, Optional<string>> => ({
+  ...position,
+  position: Optional.some(position.position)
 });
 
-const applyPositionCss = (element: SugarElement, position: PositionCss): void => {
-  const addPx = (num: number) => num + 'px';
-
-  Css.setOptions(element, {
-    position: Optional.some(position.position),
-    left: position.left.map(addPx),
-    top: position.top.map(addPx),
-    right: position.right.map(addPx),
-    bottom: position.bottom.map(addPx)
-  });
+const applyPositionCss = (element: SugarElement<HTMLElement>, position: PositionCss): void => {
+  Css.setOptions(element, toOptions(position));
 };
 
 export {