+++ modules/alloy/src/test/ts/browser/position/view/BounderToolbuttonTest.ts
@@ -10,7 +10,7 @@ import { AnchorBox, AnchorElement, AnchorLayout } from 'ephox/alloy/positioning/
 import * as Bounder from 'ephox/alloy/positioning/view/Bounder';
 
 interface TestDecisionSpec {
-  readonly label: string;
+  readonly layout: string;
   readonly x: number;
   readonly y: number;
   readonly candidateY?: number;
@@ -20,7 +20,7 @@ describe('BounderToolbuttonTest', () => {
   const check = (expected: TestDecisionSpec, preference: AnchorLayout[], anchor: AnchorBox, panel: AnchorElement, bubbles: Bubble, bounds: Bounds) => {
     const placee = SugarElement.fromTag('div');
     const actual = Bounder.attempts(placee, preference, anchor, panel, bubbles, bounds);
-    assert.equal(actual.label, expected.label, 'label');
+    assert.equal(actual.layout, expected.layout, 'layout');
     assert.equal(actual.rect.x, expected.x, 'X');
     assert.equal(actual.rect.y, expected.y, 'Y');
     if (expected.candidateY !== undefined) {
@@ -111,13 +111,13 @@ describe('BounderToolbuttonTest', () => {
 
   it('an empty input array is invalid and just returns anchor coordinates', () => {
     check({
-      label: 'none',
+      layout: 'none',
       x: 0,
       y: 0
     }, [], bounds(0, 0, 10, 10), bounds(0, 0, 50, 50), bubble(0), bounds(0, 0, 1000, 1000));
 
     check({
-      label: 'none',
+      layout: 'none',
       x: 100,
       y: 0
     }, [], bounds(100, 0, 200, 50), bounds(0, 0, 150, 25), bubble(10), bounds(0, 0, 1000, 1000));
@@ -130,7 +130,7 @@ describe('BounderToolbuttonTest', () => {
   it('southeast', () => {
     const anchor = bounds(100, 55, 10, 10);
     check({
-      label: 'layout-southeast',
+      layout: 'layout-southeast',
       x: anchor.x - 1,                            // 99
       y: anchor.bottom - 2                        // 63
     }, four, anchor, panelBox, bubb, view);
@@ -139,7 +139,7 @@ describe('BounderToolbuttonTest', () => {
   it('southwest', () => {
     const anchor = bounds(320, 55, 30, 10);
     check({
-      label: 'layout-southwest',
+      layout: 'layout-southwest',
       x: anchor.right - panelBox.width + 31,      // 281
       y: anchor.bottom - 2                        // 63
     }, four, anchor, panelBox, bubb, view);
@@ -148,7 +148,7 @@ describe('BounderToolbuttonTest', () => {
   it('northeast', () => {
     const anchor = bounds(140, 235, 10, 10);
     check({
-      label: 'layout-northeast',
+      layout: 'layout-northeast',
       x: anchor.x - 1,                            // 139
       y: anchor.y - panelBox.height + 1           // 161
     }, four, anchor, panelBox, bubb, view);
@@ -157,7 +157,7 @@ describe('BounderToolbuttonTest', () => {
   it('northwest', () => {
     const anchor = bounds(320, 235, 10, 10);
     check({
-      label: 'layout-northwest',
+      layout: 'layout-northwest',
       x: anchor.right - panelBox.width + 31,      // 261
       y: anchor.y - panelBox.height + 1           // 161
     }, four, anchor, panelBox, bubb, view);
@@ -166,7 +166,7 @@ describe('BounderToolbuttonTest', () => {
   it('all fit -> southeast because of order of preference', () => {
     const anchor = bounds(270, 100, 10, 10);
     check({
-      label: 'layout-southeast',
+      layout: 'layout-southeast',
       x: anchor.x - 1,                            // 269
       y: anchor.bottom - 2                        // 108
     }, four, anchor, panelBox, bubb, view);
@@ -175,7 +175,7 @@ describe('BounderToolbuttonTest', () => {
   it('none near top left -> best fit is southeast', () => {
     const anchor = bounds(55, 55, 10, 10);
     check({
-      label: 'layout-southeast',
+      layout: 'layout-southeast',
       x: anchor.x - 1,                            // 54
       y: anchor.bottom - 2                        // 63
     }, four, anchor, bigPanel, bubb, view);
@@ -184,7 +184,7 @@ describe('BounderToolbuttonTest', () => {
   it('none near top right -> best fit is southwest', () => {
     const anchor = bounds(350, 55, 10, 10);
     check({
-      label: 'layout-southwest',
+      layout: 'layout-southwest',
       x: anchor.right - bigPanel.width + 31,      // 316
       y: anchor.bottom - 2                        // 63
     }, four, anchor, bigPanel, bubb, view);
@@ -193,7 +193,7 @@ describe('BounderToolbuttonTest', () => {
   it('none near bottom left -> best fit is northeast', () => {
     const anchor = bounds(55, 200, 10, 10);
     check({
-      label: 'layout-northeast',
+      layout: 'layout-northeast',
       x: anchor.x - 1,                            // 54
       y: view.y,                                  // 50 - constrained within viewport
       candidateY: anchor.y - bigPanel.height + 1, // -299
@@ -203,7 +203,7 @@ describe('BounderToolbuttonTest', () => {
   it('none near bottom right -> best fit is northwest', () => {
     const anchor = bounds(350, 200, 10, 10);
     check({
-      label: 'layout-northwest',
+      layout: 'layout-northwest',
       x: anchor.right - bigPanel.width + 31,      // 316
       y: view.y,                                  // 50 - constrained within viewport
       candidateY: anchor.y - bigPanel.height + 1, // -299
@@ -214,7 +214,7 @@ describe('BounderToolbuttonTest', () => {
     // Note: The additional +1/+2 here is to account for the southeast bubble x/y offsets
     const anchor = bounds(view.right - panelBox.width + 1 - 1, view.bottom - panelBox.height - 10 + 2 - 1, 10, 10);
     check({
-      label: 'layout-southeast',
+      layout: 'layout-southeast',
       x: anchor.x - 1,                            // 299
       y: anchor.bottom - 2,                       // 194
     }, four, anchor, panelBox, bubb, view);
@@ -224,7 +224,7 @@ describe('BounderToolbuttonTest', () => {
     // Note: The additional +1/+2 here is to account for the southeast bubble x/y offsets
     const anchor = bounds(view.right - panelBox.width + 1, view.bottom - panelBox.height - 10 + 2, 10, 10);
     check({
-      label: 'layout-southeast',
+      layout: 'layout-southeast',
       x: anchor.x - 1,                            // 300
       y: anchor.bottom - 2,                       // 195
     }, four, anchor, panelBox, bubb, view);
@@ -234,7 +234,7 @@ describe('BounderToolbuttonTest', () => {
     // Note: The additional +1/+2 here is to account for the southeast bubble x/y offsets
     const anchor = bounds(view.right - panelBox.width + 1 + 1, view.bottom - panelBox.height - 10 + 2, 10, 10);
     check({
-      label: 'layout-southwest',
+      layout: 'layout-southwest',
       x: anchor.right - panelBox.width + 31,      // 243
       y: anchor.bottom - 2,                       // 195
     }, four, anchor, panelBox, bubb, view);
@@ -244,7 +244,7 @@ describe('BounderToolbuttonTest', () => {
     // Note: The additional +1/+2 here is to account for the southeast bubble x/y offsets
     const anchor = bounds(view.right - panelBox.width + 1, view.bottom - panelBox.height - 10 + 2 + 1, 10, 10);
     check({
-      label: 'layout-northeast',
+      layout: 'layout-northeast',
       x: anchor.x - 1,                            // 300
       y: anchor.y - panelBox.height + 1           // 114
     }, four, anchor, panelBox, bubb, view);
@@ -254,7 +254,7 @@ describe('BounderToolbuttonTest', () => {
     // Note: The additional +1/+2 here is to account for the southeast bubble x/y offsets
     const anchor = bounds(view.right - panelBox.width + 1 + 1, view.bottom - panelBox.height - 10 + 2 + 1, 10, 10);
     check({
-      label: 'layout-northwest',
+      layout: 'layout-northwest',
       x: anchor.right - panelBox.width + 31,      // 243
       y: anchor.y - panelBox.height + 1           // 114
     }, four, anchor, panelBox, bubb, view);
@@ -263,7 +263,7 @@ describe('BounderToolbuttonTest', () => {
   it('east', () => {
     const anchor = bounds(55, 150, 10, 10);
     check({
-      label: 'layout-east',
+      layout: 'layout-east',
       x: anchor.right - 1,                                          // 64
       y: anchor.y + (anchor.height / 2) - (panelBox.height / 2) + 1 // 118.5
     }, two, anchor, panelBox, bubb, view);
@@ -272,7 +272,7 @@ describe('BounderToolbuttonTest', () => {
   it('none near bottom left -> best fit is east (limited to bottom bounds)', () => {
     const anchor = bounds(55, 240, 10, 10);
     check({
-      label: 'layout-east',
+      layout: 'layout-east',
       x: anchor.right - 1,                        // 64
       y: view.bottom - panelBox.height            // 195 - constrained within viewport
     }, two, anchor, panelBox, bubb, view);
@@ -281,7 +281,7 @@ describe('BounderToolbuttonTest', () => {
   it('none near top left -> best fit is east (limited to top bounds)', () => {
     const anchor = bounds(55, 80, 10, 10);
     check({
-      label: 'layout-east',
+      layout: 'layout-east',
       x: anchor.right - 1,                        // 64
       y: view.y,                                  // 50 - constrained within viewport
     }, two, anchor, panelBox, bubb, view);
@@ -290,7 +290,7 @@ describe('BounderToolbuttonTest', () => {
   it('west', () => {
     const anchor = bounds(350, 150, 10, 10);
     check({
-      label: 'layout-west',
+      layout: 'layout-west',
       x: anchor.x - panelBox.width - 1,                             // 249
       y: anchor.y + (anchor.height / 2) - (panelBox.height / 2) + 1 // 118.5
     }, two, anchor, panelBox, bubb, view);
@@ -299,7 +299,7 @@ describe('BounderToolbuttonTest', () => {
   it('none near bottom right -> best fit is west (limited to bottom bounds)', () => {
     const anchor = bounds(350, 240, 10, 10);
     check({
-      label: 'layout-west',
+      layout: 'layout-west',
       x: anchor.x - panelBox.width - 1,           // 249
       y: view.bottom - panelBox.height            // 195 - constrained within viewport
     }, two, anchor, panelBox, bubb, view);
@@ -308,7 +308,7 @@ describe('BounderToolbuttonTest', () => {
   it('none near top right -> best fit is west (limited to top bounds)', () => {
     const anchor = bounds(350, 80, 10, 10);
     check({
-      label: 'layout-west',
+      layout: 'layout-west',
       x: anchor.x - panelBox.width - 1,           // 249
       y: view.y,                                  // 50 - constrained within viewport
     }, two, anchor, panelBox, bubb, view);